+++ libngpu/src/opengl/glcontext_egl.c
@@ -583,37 +583,6 @@ try_again:;
     return 0;
 }
 
-static int egl_init_external(struct glcontext *ctx, uintptr_t display, uintptr_t window, uintptr_t other)
-{
-    struct egl_priv *egl = ctx->priv_data;
-
-    egl->handle = other ? (EGLContext)other : eglGetCurrentContext();
-    if (!egl->handle) {
-        LOG(ERROR, "could not retrieve EGL context");
-        return NGPU_ERROR_EXTERNAL;
-    }
-
-    egl->display = eglGetCurrentDisplay();
-    if (!egl->display) {
-        LOG(ERROR, "could not retrieve EGL display");
-        return NGPU_ERROR_EXTERNAL;
-    }
-
-    egl->surface = eglGetCurrentSurface(EGL_DRAW);
-
-    egl->extensions = eglQueryString(egl->display, EGL_EXTENSIONS);
-    if (!egl->extensions) {
-        LOG(ERROR, "could not retrieve EGL extensions");
-        return NGPU_ERROR_EXTERNAL;
-    }
-
-    int ret = egl_probe_extensions(ctx);
-    if (ret < 0)
-        return ret;
-
-    return 0;
-}
-
 static void egl_uninit(struct glcontext *ctx)
 {
     struct egl_priv *egl = ctx->priv_data;
@@ -757,11 +726,3 @@ const struct glcontext_class ngpu_glcontext_egl_class = {
     .priv_size = sizeof(struct egl_priv),
 };
 
-const struct glcontext_class ngpu_glcontext_egl_external_class = {
-    .init = egl_init_external,
-    .make_current = egl_make_current,
-    .get_proc_address = egl_get_proc_address,
-    .get_handle = egl_get_handle,
-    .get_display = get_display,
-    .priv_size = sizeof(struct egl_priv),
-};